langgraph_starter/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
langgraph==0.2.33
langchain-core==0.3.15
langchain-openai==0.2.3
